+++ lib/components/builders/conditions/if.dart
@@ -1,5 +1,3 @@
 import 'package:pipen/components/builders/conditions/pipen_build_if.dart';
 
-class If extends PipenBuildIf {
-  const If({required super.condition, required super.child, super.def, super.key});
-}
+typedef If = PipenBuildIf;